requirements.txt
python3-picamera2
python3-flask
python3-waitress
python3-opencv